Исправил все баги и ошибки

+++ Lab1/main.py
@@ -2,6 +2,7 @@ import os
 import shutil
 import requests
 from bs4 import BeautifulSoup as BS
+HEADERS = {"User-Agent": "Mozilla/5.0"}
 
 URL = "https://yandex.ru/images/"  # ссылка на страничку html
 
@@ -9,17 +10,21 @@ URL = "https://yandex.ru/images/"  # ссылка на страничку html
 def save_image(image_url, name, i):
     """сохранение картинки в папку"""
     req = requests.get(f"https:{image_url}")
-    with open(f"dataset/{name}/{i:04d}.jpg", "wb") as saver:
+    file = os.path.join(f"dataset/{name}/{i:04d}.jpg")
+    with open(file, "wb") as saver:
         saver.write(req.content)
 
 
 def check_folder():
     """проверка существования папки"""
     try:
-        os.mkdir("dataset")
-    except:
-        shutil.rmtree("dataset")
-        os.mkdir("dataset")
+        if not os.path.isdir("dataset"):
+            os.mkdir("dataset")
+        else:
+            shutil.rmtree("dataset")
+            os.mkdir("dataset")
+    except OSError as err:
+        print(f"Возникла ошибка!!{err}")
 
 
 def get_images_url(name):
@@ -30,21 +35,23 @@ def get_images_url(name):
     """
     i = 1
     page = 0
-    request = requests.get(f"{URL}search?p={page}&text={name}&lr=51&rpt=image", headers={"User-Agent": "Mozilla/5.0"})
+    request = requests.get(f"{URL}search?p={page}&text={name}&lr=51&rpt=image", HEADERS)
     html = BS(request.content, "html.parser")
     data = []
     searcher = html.findAll("img")
-    os.mkdir(f"dataset/{name}")
-    if (i < 999):
-        for event in searcher:
-            image_url = event.get("src")
-            data.append([image_url])
-            if (i > 999):
-                page = 0
-                break
-            if (image_url != ""):
-                save_image(image_url, name, i)
-                i += 1
+    try:
+        os.mkdir(f"dataset/{name}")
+    except PermissionError:
+        print("Unable to create a directory")
+    for event in searcher:
+        image_url = event.get("src")
+        data.append([image_url])
+        if (image_url != ""):
+            save_image(image_url, name, i)
+            i += 1
+        if (i > 999):
+            page = 0
+            break
         page += 1
     print("Images save: ")
     print(data)
